requirements.txt
langchain_core==0.2.43
langchain_ollama==0.1.3
langgraph==0.3.25
lifelines==0.29.0
matplotlib
numpy 
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
Pillow>=10.0.0
plotnine==0.13.6
python_Levenshtein>=0.25.1
scikit_learn==1.6.0
scipy
seaborn==0.13.2
statsmodels==0.14.2
tkhtmlview~=0.3.1
typing_extensions==4.12.2
regex>=2021.8.3
tqdm~=4.67.1